requirements.txt
Flask==3.0.3
twilio==9.2.3
google-cloud-speech==2.27.0
python-dotenv==1.0.1
requests==2.32.3
